python/requirements.txt
PyGithub~=2.1.1
requests~=2.31.0
gitpython~=3.1.40